backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
python-multipart
httpx
web3
snowflake-connector-python
google-cloud-storage
